Bug #584 fix warnings

+++ src/tcpedit/plugins/dlt_en10mb/en10mb.c
@@ -773,7 +773,7 @@ dlt_en10mb_l2len(tcpeditdlt_t *ctx, const u_char *packet, const int pktlen)
 
     ether_type = ntohs(((eth_hdr_t*)packet)->ether_type);
     while (ether_type == ETHERTYPE_VLAN) {
-        if (pktlen < l2len + sizeof(vlan_hdr_t))
+        if (pktlen < l2len + (int)sizeof(vlan_hdr_t))
              return -1;
 
          vlan_hdr_t *vlan_hdr = (vlan_hdr_t*)(packet + l2len);
